Make bot compile'
